// src/generated/models/New_deliverablesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const New_deliverablesnew_status = {
  100000000: 'Delivered',
  100000001: 'ToBeDelivered',
  100000002: 'Delayed'
} as const;
export type New_deliverablesnew_status = keyof typeof New_deliverablesnew_status;
export const New_deliverablesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type New_deliverablesstatecode = keyof typeof New_deliverablesstatecode;
export const New_deliverablesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type New_deliverablesstatuscode = keyof typeof New_deliverablesstatuscode;

export interface New_deliverablesBase {
  importsequencenumber?: number;
  new_deliverableid: string;
  new_deliverablesid: string;
  new_notes?: string;
  new_projectcategory?: string;
  new_projectgoal?: string;
  new_projectmanager?: string;
  new_projectname: string;
  new_status?: New_deliverablesnew_status;
  new_thedeliverablesinclude?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: New_deliverablesstatecode;
  statuscode?: New_deliverablesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface New_deliverables extends New_deliverablesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  new_statusname?: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}